Add default locale (#17)
Add a default arg for the locale parameter of the Pattern constructor.
It already provides default args for position and zoom, and it makes
sense to add a default arg for locale as well, since there may be
circumstances (on the command line, for example, or for testing)
where you just quickly want to create a pattern and don't care about
localization.