Add IO and Visual modules (#19)
Add I/O and visual functionality to the patterns-ts package. This
functionality will be shared by the Pattern Explorer and the CLI
scripts for SVG generation. Most of the added code was moved from the
Pattern Explorer repository, with a few modifications: stricter type
checking, input validation of references to nodes, node definitions,
and categories, and ECMA compliant import statements. Further changes
to this repository's infrastructure include disabling of autoformatting
for JSON with comments files, adding React and React Flow dependencies,
and bumping the version to 1.0.0-alpha.5.